% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDB.installers.R
\name{AddGenome}
\alias{AddGenome}
\title{AddGenome
This function will install a Genome into your database}
\usage{
AddGenome(species, urlFasta, urlGTF, version)
}
\arguments{
\item{species}{string}

\item{urlGTF}{string, the url of the genome GTF annotation file}

\item{version}{string denoting the version name to identify the assembly+annotation pair}

\item{urlfasta}{string, the url of the genome assembly}
}
\description{
AddGenome
This function will install a Genome into your database
}
