% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDB.installers.R
\name{AddHumanGenomes}
\alias{AddHumanGenomes}
\title{AddHumanGenomes}
\usage{
if any is missing, the GRCh38 assembly and the GENCODE release 38 annotation GTF is installed with version name "GRCh38+GENCODE". Else
all the parameters are required
}
\arguments{
\item{urlFasta}{url for the genome fasta file}

\item{urlGTF}{the url of the annotation file}

\item{version}{the name of the version identifying assembly+annotation sources}
}
\description{
AddHumanGenomes
}
