% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDB.installers.R
\name{.ValidateGenomeGTF}
\alias{.ValidateGenomeGTF}
\title{.ValidateGenomeGTF
This function check that fasta genome file and GTF annotation file has the same chromosome identification code
This should not be used directly (only internal)}
\usage{
.ValidateGenomeGTF(filesGenomeGTF)
}
\arguments{
\item{fileGenomeGTF}{see \code{\link{GetGenome}}}
}
\description{
.ValidateGenomeGTF
This function check that fasta genome file and GTF annotation file has the same chromosome identification code
This should not be used directly (only internal)
}
