% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDB.installers.R
\name{GetGenome}
\alias{GetGenome}
\title{GetGenome
return a list with the slots fasta and gtf with the full path to the genomes and annotation files}
\usage{
GetGenome(specie, version)
}
\arguments{
\item{specie}{the stored specie (Human" etc..)}

\item{version}{"the assembly+annotation verion/code}
}
\description{
GetGenome
return a list with the slots fasta and gtf with the full path to the genomes and annotation files
}
