% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDB.installers.R
\name{GetGenomesInDB}
\alias{GetGenomesInDB}
\title{GetGenomesInDB
return the available genomes in the Database}
\usage{
genome.list <- GetGenomesInDB()
}
\description{
GetGenomesInDB
return the available genomes in the Database
}
