% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDB.installers.R
\name{GetGenome}
\alias{GetGenome}
\title{GetGenome
return a list with the slots fasta and gtf with the full path to the genomes and annotation files}
\usage{
GetGenome(species, version)
}
\arguments{
\item{version}{"the assembly+annotation version/code}

\item{specie}{the stored specie (Human" etc..)}
}
\description{
GetGenome
return a list with the slots fasta and gtf with the full path to the genomes and annotation files
}
\examples{
\dontrun{
genome.annot <- GetGenome(species = "human", version = "GRCh38+GENECODE")
}
}
