% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeDB.installers.R
\name{SetDBdirectory}
\alias{SetDBdirectory}
\title{SetDBdirectory}
\usage{
SetDBdirectory("my/preferred/path")
}
\arguments{
\item{main}{directory path name}
}
\description{
Set GenomeDB directory for the RSAP family.
It will create a directory .../main/GenomesDB, where main is "my/favorite/directory"
This function should be run only once
}
